Equipment stats changed from int to float

+++ Assets/Internal assets/Scripts/Other/IModifiers.cs
@@ -2,6 +2,6 @@ namespace Other
 {
     public interface IModifiers
     {
-        void AddValue(ref int baseValue);
+        void AddValue(ref float baseValue);
     }
 }
